Kobe/requirements.lock.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
curl_cffi==0.14.0b2
distro==1.9.0
duckduckgo_search==5.3.1
fake-useragent==2.2.0
fastapi==0.119.0
frozenlist==1.8.0
griffe==1.14.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
httpx-socks==0.10.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
lxml_html_clean==0.4.2
markdown-it-py==4.0.0
mcp==1.19.0
mdurl==0.1.2
msgspec==0.19.0
multidict==6.7.0
openai==2.5.0
openai-agents==0.4.0
orjson==3.11.4
pdfminer.six==20240706
pillow==11.3.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
python-socks==2.7.2
python-telegram-bot==22.3
pytz==2025.2
pywin32==311
PyYAML==6.0.3
redis==7.0.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.2
starlette==0.48.0
tornado==6.5.2
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
yarl==1.22.0
